Refactor App.js by splitting components into separate files and organizing them into a components folder. Updated imports and exports for consistency using default exports.
